+++ packages/stream-model-instance-document/src/model-instance-document.ts
